Fix the type of "custom_event_rules"
Switched it to "any" to allow using different structs in the rules and avoid terrafom throwing an error
while trying to do a duct typing of all elements.
In order to keep the variable type static, we added a validation block. It checks that "detail" and "detail-type" is always set